Use decimal instead of float, to get consistent rounding

+++ custom_components/powercalc/helpers.py
@@ -1,9 +1,10 @@
+from decimal import Decimal
 from typing import Union
 
 from homeassistant.helpers.template import Template
 
 
-async def evaluate_power(power: Union[Template, float]) -> float:
+async def evaluate_power(power: Union[Template, Decimal]) -> Decimal:
     """When power is a template render it."""
 
     if isinstance(power, Template):
